rhema_search/src/requirements.txt
anyio==3.6.2
Authlib==1.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.1
decorator==5.1.1
fastapi==0.92.0
h11==0.14.0
idna==3.4
loguru==0.6.0
pycparser==2.21
pydantic==1.10.5
requests==2.28.2
sniffio==1.3.0
starlette==0.25.0
tqdm==4.64.1
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.20.0
validators==0.19.0
weaviate-client==3.12.0
